+++ src/main/java/webserver/RequestHandler.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import util.RequestParser;
+import webserver.request.RequestParser;
 import util.ResponseBuilder;
 
 public class RequestHandler implements Runnable {
